Move report processors into separate package
This gives a nicer separate between the generic pipeline logic, and the
specific processors that will be available to add via configuration
files.  (Some of the beefier processors, which have their own large
dependencies on other Go packages, will get their own standalone
packages.)
